requirements.txt
streamlit
langchain
pinecone-client
sentence-transformers
python-dotenv
